Move comms state into core.entities
