+++ packages/reducers/src/index.ts
@@ -1,5 +1,4 @@
 import app from "./app";
-import comms from "./comms";
 import core from "./core";
 
-export { comms, app, core };
+export { app, core };
